AIS-1198 -refactor code
