+++ ai-services/internal/pkg/application/common/backup/digitize.go
@@ -1,10 +1,14 @@
-package backup
+package common
 
 import (
+	"archive/tar"
+	"compress/gzip"
 	"encoding/json"
 	"fmt"
+	"io"
 	"os"
 	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/go-resty/resty/v2"
@@ -12,9 +16,14 @@ import (
 )
 
 const (
-	exportAllRecordsLimit = "-1"
+	ExportAllRecordsLimit = "-1"
+	defaultFilePermission = 0o644
+	defaultDirPermission  = 0o755
+	bytesPerKB            = 1024
+	bytesPerMB            = bytesPerKB * 1024
 )
 
+// DigitizeExportResponse represents the response from the digitize export API.
 type DigitizeExportResponse struct {
 	Status          string                   `json:"status"`
 	Data            DigitizeImportExportData `json:"data"`
@@ -24,22 +33,26 @@ type DigitizeExportResponse struct {
 	Pagination      DigitizeExportPagination `json:"pagination"`
 }
 
+// DigitizeImportExportData contains the jobs and documents data.
 type DigitizeImportExportData struct {
 	Jobs      []map[string]interface{} `json:"jobs"`
 	Documents []map[string]interface{} `json:"documents"`
 }
 
+// DigitizeExportSummary contains summary information for the export.
 type DigitizeExportSummary struct {
 	Jobs      DigitizeExportEntitySummary `json:"jobs"`
 	Documents DigitizeExportEntitySummary `json:"documents"`
 }
 
+// DigitizeExportEntitySummary contains summary for a specific entity type.
 type DigitizeExportEntitySummary struct {
 	TotalExported int `json:"total_exported"`
 	Completed     int `json:"completed"`
 	Failed        int `json:"failed"`
 }
 
+// DigitizeExportPagination contains pagination information.
 type DigitizeExportPagination struct {
 	Limit           int  `json:"limit"`
 	Offset          int  `json:"offset"`
@@ -55,8 +68,10 @@ type DigitizeBackupClient struct {
 
 // NewDigitizeBackupClient creates a new digitize backup client.
 func NewDigitizeBackupClient(serviceURL string) *DigitizeBackupClient {
+	client := resty.New().SetBaseURL(serviceURL)
+
 	return &DigitizeBackupClient{
-		client: resty.New().SetBaseURL(serviceURL),
+		client: client,
 	}
 }
 
@@ -68,7 +83,7 @@ func (c *DigitizeBackupClient) CallExportAPI() (*DigitizeExportResponse, error) 
 
 	logger.Infof("Sending export request to: /v1/export?limit=-1\n", 0)
 	resp, err := c.client.R().
-		SetQueryParam("limit", exportAllRecordsLimit).
+		SetQueryParam("limit", ExportAllRecordsLimit).
 		SetResult(&exportResponse).
 		Get("/v1/export")
 
@@ -83,6 +98,7 @@ func (c *DigitizeBackupClient) CallExportAPI() (*DigitizeExportResponse, error) 
 	return &exportResponse, nil
 }
 
+// CreateDigitizeBackupArchive creates a backup archive from the export response.
 func CreateDigitizeBackupArchive(backupFile string, exportResponse *DigitizeExportResponse) error {
 	logger.Infof("Creating digitize backup archive...\n", 0)
 
@@ -181,6 +197,136 @@ func writeJSONFile(path string, data interface{}) error {
 	}
 
 	return nil
+}
+
+// CreateTarGzArchive creates a tar.gz archive from a source directory.
+// It includes the specified entries (files or directories) in the archive.
+func CreateTarGzArchive(sourceDir, targetFile string, entries []string) error {
+	file, err := os.Create(targetFile)
+	if err != nil {
+		return fmt.Errorf("failed to create backup archive: %w", err)
+	}
+	defer func() {
+		_ = file.Close()
+	}()
+
+	gzipWriter := gzip.NewWriter(file)
+	defer func() {
+		_ = gzipWriter.Close()
+	}()
+
+	tarWriter := tar.NewWriter(gzipWriter)
+	defer func() {
+		_ = tarWriter.Close()
+	}()
+
+	for _, entry := range entries {
+		fullPath := filepath.Join(sourceDir, entry)
+		if err := addPathToTar(tarWriter, sourceDir, fullPath); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// addPathToTar adds a file or directory to the tar archive.
+func addPathToTar(tarWriter *tar.Writer, baseDir, path string) error {
+	info, err := os.Stat(path)
+	if err != nil {
+		return fmt.Errorf("failed to stat path %s: %w", path, err)
+	}
+
+	relPath, err := filepath.Rel(baseDir, path)
+	if err != nil {
+		return fmt.Errorf("failed to determine relative path for %s: %w", path, err)
+	}
+
+	header, err := tar.FileInfoHeader(info, "")
+	if err != nil {
+		return fmt.Errorf("failed to create tar header for %s: %w", path, err)
+	}
+	header.Name = filepath.ToSlash(relPath)
+
+	if info.IsDir() && header.Name[len(header.Name)-1] != '/' {
+		header.Name += "/"
+	}
+
+	if err := tarWriter.WriteHeader(header); err != nil {
+		return fmt.Errorf("failed to write tar header for %s: %w", path, err)
+	}
+
+	if info.IsDir() {
+		return addDirectoryToTar(tarWriter, baseDir, path)
+	}
+
+	return addFileToTar(tarWriter, path)
+}
+
+// addDirectoryToTar recursively adds a directory's contents to the tar archive.
+func addDirectoryToTar(tarWriter *tar.Writer, baseDir, path string) error {
+	entries, err := os.ReadDir(path)
+	if err != nil {
+		return fmt.Errorf("failed to read directory %s: %w", path, err)
+	}
+
+	for _, entry := range entries {
+		if err := addPathToTar(tarWriter, baseDir, filepath.Join(path, entry.Name())); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// addFileToTar adds a file's contents to the tar archive.
+func addFileToTar(tarWriter *tar.Writer, path string) error {
+	file, err := os.Open(path)
+	if err != nil {
+		return fmt.Errorf("failed to open file %s: %w", path, err)
+	}
+	defer func() {
+		_ = file.Close()
+	}()
+
+	if _, err := io.Copy(tarWriter, file); err != nil {
+		return fmt.Errorf("failed to write file contents for %s: %w", path, err)
+	}
+
+	return nil
+}
+
+// LogArchiveSize logs the size of the created archive file.
+func LogArchiveSize(backupFile string) {
+	fileInfo, err := os.Stat(backupFile)
+	if err == nil {
+		sizeMB := float64(fileInfo.Size()) / bytesPerMB
+		logger.Infof("✓ Tar archive created: %s (%.2f MB)\n", backupFile, sizeMB, 0)
+
+		return
+	}
+
+	logger.Infof("✓ Tar archive created: %s\n", backupFile, 0)
+}
+
+func GetBackupFile(backupFile string, appName string) (string, error) {
+	if backupFile == "" {
+		timestamp := time.Now().Format("20060102_150405")
+		backupFile = fmt.Sprintf("%s_digitize_backup_%s.tar.gz", appName, timestamp)
+	}
+
+	// Ensure .tar.gz extension
+	if !strings.HasSuffix(backupFile, ".tar.gz") {
+		backupFile += ".tar.gz"
+	}
+
+	// Get absolute path for backup file
+	absBackupFile, err := filepath.Abs(backupFile)
+	if err != nil {
+		return "", fmt.Errorf("failed to get absolute path for backup file: %w", err)
+	}
+
+	return absBackupFile, nil
 }
 
 // Made with Bob
